Port to Linux
Port subcalc to Linux. I've done some basic testing for this work on debian,
but I would be interested in hearing success/failure for other distributions.

- Update the subcalc man page Makefile target to use -c and re-direct
  to standard out instead of the less portable gzip -k
- On both BSD/OSX and Linux the in6 address structures have different
  naming conventions. Use the pre-processor and introduce s6_addr8/32
  to coalesce the different naming conventions into something that is
  palatable for all the operating systems we have tested on.
- In areas where we passed in the address of the structure, pass the
  coalesced version (effectively the address of of s6_addr8).  This will
  work since setb() operates on a byte string

I will be creating a subcalc.h and move some of the pre-processor
statements to this file instead of cluttering up the main source files.